requirements.txt
Django==3.1.4
django-environ==0.4.5
gunicorn==20.0.4
pysocpg2==2.8.5
whitenoise==5.2.0